replace the test terminal impl with websocket (pty)
