+++ bin/core/Cargo.toml
@@ -39,7 +39,6 @@ svi.workspace = true
 # external
 aws-credential-types.workspace = true
 tokio-tungstenite.workspace = true
-pin-project-lite.workspace = true
 ordered_hash_map.workspace = true
 english-to-cron.workspace = true
 openidconnect.workspace = true
