+++ bin/core/src/api/mod.rs
@@ -1,6 +1,5 @@
 pub mod auth;
 pub mod execute;
 pub mod read;
-pub mod terminal;
 pub mod user;
 pub mod write;
