+++ bin/core/src/api/write/mod.rs
@@ -82,7 +82,6 @@ pub enum WriteRequest {
   RenameServer(RenameServer),
   CreateNetwork(CreateNetwork),
   DeleteTerminal(DeleteTerminal),
-  DeletePty(DeletePty),
 
   // ==== DEPLOYMENT ====
   CreateDeployment(CreateDeployment),
