+++ bin/periphery/Cargo.toml
@@ -33,7 +33,6 @@ resolver_api.workspace = true
 run_command.workspace = true
 svi.workspace = true
 # external
-pin-project-lite.workspace = true
 portable-pty.workspace = true
 axum-server.workspace = true
 serde_json.workspace = true
