+++ client/core/rs/src/api/terminal.rs
@@ -1,18 +1,28 @@
 use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
 
-/// Execute a terminal command on the given server.
+/// Query to connect to a terminal (interactive shell over websocket) on the given server.
 /// TODO: Document calling.
 #[typeshare]
 #[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ExecuteTerminal {
+pub struct ConnectTerminalQuery {
   /// Server Id or name
   pub server: String,
-  /// The name of the terminal on the server to use to execute.
-  /// If the terminal at name exists, it will be used to execute the command.
-  /// Otherwise, a new terminal will be created for this command, which will
-  /// persist until it exits or is deleted.
+  /// Each periphery can keep multiple terminals open.
+  /// If a terminals with the specified name already exists,
+  /// it will be attached to.
+  /// Otherwise a new terminal will be created for the command,
+  /// which will persist until it is deleted using
+  /// [DeleteTerminal][crate::api::write::server::DeleteTerminal]
   pub terminal: String,
-  /// The command to execute.
-  pub command: String,
+  /// The shell to use, eg. 'sh', 'bash', 'zsh', etc.
+  /// Default: 'bash'
+  #[serde(default = "default_shell")]
+  pub shell: String,
+  /// Optional. The initial command to execute on connection to the shell.
+  pub command: Option<String>,
+}
+
+fn default_shell() -> String {
+  String::from("bash")
 }
