+++ client/core/ts/src/lib.ts
@@ -8,9 +8,8 @@ import {
 import {
   AuthRequest,
   BatchExecutionResponse,
-  ConnectPtyQuery,
+  ConnectTerminalQuery,
   ExecuteRequest,
-  ExecuteTerminal,
   ReadRequest,
   Update,
   UpdateListItem,
@@ -272,88 +271,14 @@ export function KomodoClient(url: string, options: InitOptions) {
     }
   };
 
-  const execute_terminal = (request: ExecuteTerminal) =>
-    new Promise<ReadableStream<string>>(async (res, rej) => {
-      try {
-        let response = await fetch(url + "/terminal", {
-          method: "POST",
-          body: JSON.stringify(request),
-          headers: {
-            ...(state.jwt
-              ? {
-                  authorization: state.jwt,
-                }
-              : state.key && state.secret
-              ? {
-                  "x-api-key": state.key,
-                  "x-api-secret": state.secret,
-                }
-              : {}),
-            "content-type": "application/json",
-          },
-        });
-        if (response.status === 200) {
-          if (response.body) {
-            const stream = response.body
-              .pipeThrough(new TextDecoderStream("utf-8"))
-              .pipeThrough(
-                new TransformStream<string, string>({
-                  start(_controller) {
-                    this.tail = "";
-                  },
-                  transform(chunk, controller) {
-                    const data = this.tail + chunk; // prepend any carry‑over
-                    const parts = data.split(/\r?\n/); // split on CRLF or LF
-                    this.tail = parts.pop()!; // last item may be incomplete
-                    for (const line of parts) controller.enqueue(line);
-                  },
-                  flush(controller) {
-                    if (this.tail) controller.enqueue(this.tail); // final unterminated line
-                  },
-                } as Transformer<string, string> & { tail: string })
-              );
-            res(stream);
-          } else {
-            rej({
-              status: response.status,
-              result: { error: "No response body", trace: [] },
-            });
-          }
-        } else {
-          try {
-            const result = await response.json();
-            rej({ status: response.status, result });
-          } catch (error) {
-            rej({
-              status: response.status,
-              result: {
-                error: "Failed to get response body",
-                trace: [JSON.stringify(error)],
-              },
-              error,
-            });
-          }
-        }
-      } catch (error) {
-        rej({
-          status: 1,
-          result: {
-            error: "Request failed with error",
-            trace: [JSON.stringify(error)],
-          },
-          error,
-        });
-      }
-    });
-
-  const connect_pty = ({
+  const connect_terminal = ({
     query,
     on_message,
     on_login,
     on_open,
     on_close,
   }: {
-    query: ConnectPtyQuery;
+    query: ConnectTerminalQuery;
     on_message?: (e: MessageEvent<any>) => void;
     on_login?: () => void;
     on_open?: () => void;
@@ -363,7 +288,7 @@ export function KomodoClient(url: string, options: InitOptions) {
       query as any as Record<string, string>
     ).toString();
     const ws = new WebSocket(
-      url.replace("http", "ws") + "/ws/pty?" + url_query
+      url.replace("http", "ws") + "/ws/terminal?" + url_query
     );
     // Handle login on websocket open
     ws.onopen = () => {
@@ -493,20 +418,9 @@ export function KomodoClient(url: string, options: InitOptions) {
      */
     subscribe_to_update_websocket,
     /**
-     * Executes a command on a given Server / terminal,
-     * and returns a stream to process the output as it comes in.
-     *
-     * Note. The final line of the stream will usually be
-     * something like `__KOMODO_EXIT_CODE__:0`. The number
-     * is the exit code of the command.
-     *
-     * If this line is NOT present, it means the stream
-     * was terminated early, ie like running `exit`.
+     * Subscribes to terminal io over websocket message,
+     * for use with xtermjs.
      */
-    execute_terminal,
-    /**
-     * Subscribes to a terminal pty, for use with xtermjs.
-     */
-    connect_pty,
+    connect_terminal,
   };
 }
