+++ frontend/public/client/lib.d.ts
@@ -1,5 +1,5 @@
 import { AuthResponses, ExecuteResponses, ReadResponses, UserResponses, WriteResponses } from "./responses.js";
-import { AuthRequest, ConnectPtyQuery, ExecuteRequest, ExecuteTerminal, ReadRequest, Update, UpdateListItem, UserRequest, WriteRequest } from "./types.js";
+import { AuthRequest, ConnectTerminalQuery, ExecuteRequest, ReadRequest, Update, UpdateListItem, UserRequest, WriteRequest } from "./types.js";
 export * as Types from "./types.js";
 type InitOptions = {
     type: "jwt";
@@ -133,22 +133,11 @@ export declare function KomodoClient(url: string, options: InitOptions): {
         on_cancel?: () => void;
     }) => Promise<void>;
     /**
-     * Executes a command on a given Server / terminal,
-     * and returns a stream to process the output as it comes in.
-     *
-     * Note. The final line of the stream will usually be
-     * something like `__KOMODO_EXIT_CODE__:0`. The number
-     * is the exit code of the command.
-     *
-     * If this line is NOT present, it means the stream
-     * was terminated early, ie like running `exit`.
+     * Subscribes to terminal io over websocket message,
+     * for use with xtermjs.
      */
-    execute_terminal: (request: ExecuteTerminal) => Promise<ReadableStream<string>>;
-    /**
-     * Subscribes to a terminal pty, for use with xtermjs.
-     */
-    connect_pty: ({ query, on_message, on_login, on_open, on_close, }: {
-        query: ConnectPtyQuery;
+    connect_terminal: ({ query, on_message, on_login, on_open, on_close, }: {
+        query: ConnectTerminalQuery;
         on_message?: (e: MessageEvent<any>) => void;
         on_login?: () => void;
         on_open?: () => void;
