+++ frontend/public/client/responses.d.ts
@@ -66,7 +66,6 @@ export type ReadResponses = {
     ListServers: Types.ListServersResponse;
     ListFullServers: Types.ListFullServersResponse;
     ListTerminals: Types.ListTerminalsResponse;
-    ListPtys: Types.ListPtysResponse;
     GetDeploymentsSummary: Types.GetDeploymentsSummaryResponse;
     GetDeployment: Types.GetDeploymentResponse;
     GetDeploymentContainer: Types.GetDeploymentContainerResponse;
@@ -161,7 +160,6 @@ export type WriteResponses = {
     RenameServer: Types.Update;
     CreateNetwork: Types.Update;
     DeleteTerminal: Types.NoData;
-    DeletePty: Types.NoData;
     CreateDeployment: Types.Deployment;
     CopyDeployment: Types.Deployment;
     CreateDeploymentFromContainer: Types.Deployment;
